ref(tags): Make an unique file for handling route for photo tags
